chore: refactor and add comments
