+++ packages/extension/src/shared/fetcher.ts
@@ -1,0 +1,6 @@
+/** generic json fetcher */
+
+export const fetcher = async (url: string) => {
+  const response = await fetch(url)
+  return await response.json()
+}
