+++ packages/extension/src/ui/features/accountTokens/tokens.service.ts
@@ -1,12 +1,9 @@
 import { BigNumber, BigNumberish } from "@ethersproject/bignumber"
 import { utils } from "ethers"
-import { formatUnits } from "ethers/lib/utils"
-import { useMemo } from "react"
 import { Abi, Contract, number, shortString, uint256 } from "starknet"
-import useSWR from "swr"
 
 import parsedErc20Abi from "../../../abis/ERC20.json"
-import { UniqueToken, getFeeToken } from "../../../shared/token"
+import { getFeeToken } from "../../../shared/token"
 import { Account } from "../accounts/Account"
 import { TokenDetails, TokenDetailsWithBalance } from "./tokens.state"
 
@@ -119,203 +116,4 @@ export const fetchFeeTokenBalance = async (
     return BigNumber.from(0)
   }
   return fetchTokenBalance(token.address, account)
-}
-
-export interface ApiTokenDetails {
-  id: number
-  address: string
-  pricingId: number
-}
-
-export interface ApiTokenDataResponse {
-  tokens: ApiTokenDetails[]
-}
-
-export interface ApiPriceDetails {
-  pricingId: number
-  ethValue: number
-  ccyValue: number
-  ethDayChange: number
-  ccyDayChange: number
-}
-
-export interface ApiPriceDataResponse {
-  prices: ApiPriceDetails[]
-}
-
-/** generic SWR json fetcher */
-
-export const fetcher = async (url: string) => {
-  const response = await fetch(url)
-  return await response.json()
-}
-
-export const lookupTokenPriceDetails = ({
-  token,
-  pricesData,
-  tokenData,
-}: {
-  token: UniqueToken
-  pricesData: ApiPriceDataResponse
-  tokenData: ApiTokenDataResponse
-}) => {
-  /** find token from tokenData by matching address */
-  const tokenInPriceData = tokenData.tokens.find(
-    ({ address }) => address.toLowerCase() === token.address.toLowerCase(),
-  )
-  if (tokenInPriceData) {
-    /** find token price details from pricesData by matching priceId */
-    const priceDetails = pricesData.prices.find(
-      ({ pricingId }) => pricingId === tokenInPriceData.pricingId,
-    )
-    return priceDetails
-  }
-}
-
-export const ARGENT_API_TOKENS_PRICES_URL = `${process.env.REACT_APP_ARGENT_API_BASE_URL}/tokens/prices?chain=starknet`
-export const ARGENT_API_TOKENS_INFO_URL = `${process.env.REACT_APP_ARGENT_API_BASE_URL}/tokens/info?chain=starknet`
-
-/** TODO: implement currency? */
-export const usePriceAndTokenData = () => {
-  const { data: pricesData } = useSWR<ApiPriceDataResponse>(
-    `${ARGENT_API_TOKENS_PRICES_URL}`,
-    fetcher,
-    {
-      refreshInterval: 60 * 1000 /** 60 seconds */,
-    },
-  )
-  const { data: tokenData } = useSWR<ApiTokenDataResponse>(
-    `${ARGENT_API_TOKENS_INFO_URL}`,
-    fetcher,
-    {
-      refreshInterval: 5 * 60 * 1000 /** 5 minutes */,
-    },
-  )
-  return {
-    pricesData,
-    tokenData,
-  }
-}
-
-export const useTokenPriceDetails = (
-  token: TokenDetails | TokenDetailsWithBalance,
-  usePriceAndTokenDataImpl = usePriceAndTokenData,
-) => {
-  const { pricesData, tokenData } = usePriceAndTokenDataImpl()
-  return useMemo(() => {
-    if (!pricesData || !tokenData) {
-      return
-    }
-    return lookupTokenPriceDetails({
-      token,
-      pricesData,
-      tokenData,
-    })
-  }, [token, pricesData, tokenData])
-}
-
-export const useTokenBalanceToCurrencyValue = (
-  token: TokenDetailsWithBalance,
-  usePriceAndTokenDataImpl = usePriceAndTokenData,
-) => {
-  const priceDetails = useTokenPriceDetails(token, usePriceAndTokenDataImpl)
-  return useMemo(() => {
-    if (!token || !priceDetails || !token.balance || !token.decimals) {
-      return
-    }
-    const currencyValue = convertTokenAmountToCurrencyValue({
-      amount: token.balance,
-      decimals: token.decimals,
-      unitCurrencyValue: priceDetails.ccyValue,
-    })
-    return currencyValue
-  }, [priceDetails, token])
-}
-
-export const useSumTokenBalancesToCurrencyValue = (
-  tokens: TokenDetailsWithBalance[],
-  usePriceAndTokenDataImpl = usePriceAndTokenData,
-) => {
-  const { pricesData, tokenData } = usePriceAndTokenDataImpl()
-  return useMemo(() => {
-    if (!tokens || !pricesData || !tokenData) {
-      return
-    }
-    let sumTokenBalance = 0
-    tokens.forEach((token) => {
-      const priceDetails = lookupTokenPriceDetails({
-        token,
-        pricesData,
-        tokenData,
-      })
-      if (!priceDetails || !token.balance || !token.decimals) {
-        // missing data - don't add it
-      } else {
-        const currencyValue = convertTokenAmountToCurrencyValue({
-          amount: token.balance,
-          decimals: token.decimals,
-          unitCurrencyValue: priceDetails.ccyValue,
-        })
-        if (currencyValue) {
-          sumTokenBalance += Number(currencyValue)
-        }
-      }
-    })
-    /** convert to string for consistency with `convertTokenAmountToCurrencyValue()` */
-    return String(sumTokenBalance)
-  }, [pricesData, tokenData, tokens])
-}
-
-export const countDecimals = (value: number | string) => {
-  const numValue = Number(value)
-  /** check for whole number with no decimals */
-  if (Math.floor(numValue) === numValue) {
-    return 0
-  }
-  /** count decimals after conversion to string e.g. '12.34' */
-  return numValue.toString().split(".")[1].length || 0
-}
-
-export const convertTokenAmountToCurrencyValue = ({
-  amount,
-  decimals,
-  unitCurrencyValue,
-}: {
-  amount: BigNumberish
-  decimals: BigNumberish
-  unitCurrencyValue: number | string
-  unformatted?: boolean
-}) => {
-  /**
-   * BigNumber is only for integers, it does not support floating-point or fixed-point math
-   * @see https://github.com/ethers-io/ethers.js/issues/488#issuecomment-481944450
-   */
-
-  const decimalsNumber = Number(decimals)
-  const unitCurrencyValueNumber = Number(unitCurrencyValue)
-
-  /** determine what we need to multiply by to make price into an integer */
-  const priceDecimals = countDecimals(unitCurrencyValueNumber)
-  const priceToIntegerMultiplier = Math.pow(10, priceDecimals)
-
-  /** Math.round due to loss of precision */
-  const integerPrince = BigNumber.from(
-    Math.round(unitCurrencyValueNumber * priceToIntegerMultiplier),
-  )
-
-  /** Multiply the integer price by balance, then divide down by the multiplier from above */
-  const priceWithDecimals = integerPrince
-    .mul(amount)
-    .div(priceToIntegerMultiplier)
-
-  /** Convert down using decimals */
-  const convertedPrice = formatUnits(priceWithDecimals, decimalsNumber)
-  return convertedPrice
-}
-
-export const prettifyCurrencyValue = (currencyValue: string | number) => {
-  const prettyValue = Number(currencyValue).toFixed(2)
-  /** TODO: implement currency? thousands separators etc.? */
-  const symbol = "$"
-  return `${symbol}${prettyValue}`
 }
